Clock correction for old dynamic compilation core
For the old dynamic compilation core, use fixed Cycles (CYCLE_MULT_DEFAULT), and multipliers_override hack is applied to the timer 2

+++ psxhw.h
@@ -76,8 +76,8 @@ void psxHwReset();
 u8   psxHwRead8 (u32 add);
 u16  psxHwRead16(u32 add);
 u32  psxHwRead32(u32 add);
-void psxHwWrite8 (u32 add, u8  value);
-void psxHwWrite16(u32 add, u16 value);
+void psxHwWrite8 (u32 add, u32 value);
+void psxHwWrite16(u32 add, u32 value);
 void psxHwWrite32(u32 add, u32 value);
 int psxHwFreeze(gzFile f, int Mode);
